fix test_timeline_deletion_with_files_stuck_in_upload_queue by allowing deletions of non existinf files in LocalFs remote storage (this is real s3 behavior)
